Feat:Polaroid 뷰 컴포넌트 역할 별 분리
