+++ Modi-frontend/src/components/HomePage/Diary/Photo/PhotoDiary.module.css
