+++ Modi-frontend/src/components/HomePage/Diary/Polaroid/PolaroidDiary.module.css
